+++ bzt/engine.py
@@ -15,26 +15,26 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 """
-import sys
 import copy
+import datetime
 import hashlib
 import json
 import logging
 import os
 import shutil
+import sys
 import time
 import traceback
-import yaml
 from abc import abstractmethod
 from collections import namedtuple, defaultdict
 from distutils.version import LooseVersion
 from json import encoder
+
+import yaml
 from yaml.representer import SafeRepresenter
 
-import datetime
-
 import bzt
-from bzt import ManualShutdown, NormalShutdown, get_configs_dir
+from bzt import ManualShutdown, get_configs_dir
 from bzt.six import build_opener, install_opener, urlopen, numeric_types, iteritems
 from bzt.six import string_types, text_type, PY2, UserDict, parse, ProxyHandler, etree, reraise
 from bzt.utils import PIPE, shell_exec, get_full_path, ExceptionalDownloader, get_uniq_name
@@ -118,7 +118,7 @@ class Engine(object):
             self.config.dump()
 
         except BaseException as exc:
-            self.stopping_reason = exc if not self.stopping_reason else self.stopping_reason
+            self.stopping_reason = exc
             raise
 
     def _startup(self):
@@ -135,18 +135,25 @@ class Engine(object):
         calls `shutdown` in any case
         """
         self.log.info("Starting...")
+        exc_info = None
         try:
             self._startup()
             self._wait()
-        except NormalShutdown as exc:
-            self.log.debug("Normal shutdown called: %s", traceback.format_exc())
-            self.stopping_reason = exc if not self.stopping_reason else self.stopping_reason
         except BaseException as exc:
-            self.stopping_reason = exc if not self.stopping_reason else self.stopping_reason
-            raise
+            self.stopping_reason = exc
+            exc_info = sys.exc_info()
         finally:
             self.log.warning("Please wait for graceful shutdown...")
-            self._shutdown()
+            try:
+                self._shutdown()
+            except BaseException as exc:
+                if not self.stopping_reason:
+                    self.stopping_reason = exc
+                if not exc_info:
+                    exc_info = sys.exc_info()
+
+        if exc_info:
+            reraise(exc_info)
 
     def _check_modules_list(self):
         finished = False
@@ -189,9 +196,8 @@ class Engine(object):
             try:
                 if module in self.started:
                     module.shutdown()
-            except BaseException as exc:
-                self.log.error("Error while shutting down: %s", traceback.format_exc())
-                self.stopping_reason = exc if not self.stopping_reason else self.stopping_reason
+            except BaseException:
+                self.log.debug("Error while shutting down: %s", traceback.format_exc())
                 if not exc_info:
                     exc_info = sys.exc_info()
 
@@ -205,7 +211,7 @@ class Engine(object):
         """
         self.log.info("Post-processing...")
         # :type exception: BaseException
-        exception = None
+        exc_info = None
         modules = [self.provisioning, self.aggregator] + self.reporters + self.services
         for module in modules:
             if module in self.prepared:
@@ -213,19 +219,17 @@ class Engine(object):
                     module.post_process()
                 except BaseException as exc:
                     if isinstance(exc, KeyboardInterrupt):
-                        self.log.error("Shutdown: %s", exc)
+                        self.log.debug("Shutdown: %s", exc)
                     else:
-                        self.log.error("Error while post-processing: %s", traceback.format_exc())
-                    if not exception:
-                        exception = exc
+                        self.log.debug("Error while post-processing: %s", exc)
+                    if not self.stopping_reason:
+                        self.stopping_reason = exc
+                    if not exc_info:
+                        exc_info = sys.exc_info()
         self.config.dump()
 
-        if exception:
-            self.log.debug("Exception in post-process: %s", exception)
-            self.stopping_reason = exception if not self.stopping_reason else self.stopping_reason
-            if not isinstance(exception, KeyboardInterrupt):
-                self.log.warning("Failed post-processing")
-            raise exception
+        if exc_info:
+            reraise(exc_info)
 
     def create_artifact(self, prefix, suffix):
         """
@@ -604,8 +608,7 @@ class Configuration(BetterDict):
         if fmt == self.JSON:
             fds.write(to_json(self))
         elif fmt == self.YAML:
-            yml = yaml.dump(self, default_flow_style=False,
-                            explicit_start=True, canonical=False)
+            yml = yaml.dump(self, default_flow_style=False, explicit_start=True, canonical=False, allow_unicode=True)
             fds.write(yml)
         else:
             raise ValueError("Unknown dump format: %s" % fmt)
@@ -933,11 +936,15 @@ class ScenarioExecutor(EngineModule):
 
         environ = BetterDict()
         environ.merge(dict(os.environ))
+
         if aliases:
             environ["HOSTALIASES"] = hosts_file
         if env is not None:
             environ.merge(env)
+
         environ.merge({"TAURUS_ARTIFACTS_DIR": self.engine.artifacts_dir})
+
+        environ = {key: environ[key] for key in environ.keys() if environ[key] is not None}
 
         return shell_exec(args, cwd=cwd, stdout=stdout, stderr=stderr, stdin=stdin, shell=shell, env=environ)