+++ bzt/jmx2yaml.py
@@ -14,6 +14,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 """
+import codecs
 import itertools
 import logging
 import os
@@ -382,7 +383,7 @@ class JMXasDict(JMX):
         protocol = "http" if not url_info.protocol else url_info.protocol
         if url_info.domain:
             return protocol + "://" + url_info.domain + port + path
-        return ""
+        return path
 
     def _extract_url_info(self, element):
         """
@@ -1254,7 +1255,7 @@ class JMX2YAML(object):
             path = os.path.join(additional_files_dir, filename)
             self.log.info("Writing additional file: %s", path)
             content = self.converter.dialect.additional_files[filename]
-            with open(path, 'w') as f:
+            with codecs.open(path, 'w', encoding='utf-8') as f:
                 f.write(content)
 
         self.log.info("Done processing, result saved in %s", file_name)