+++ bzt/modules/blazemeter.py
@@ -101,7 +101,7 @@ class BlazeMeterUploader(Reporter, AggregatorListener, MonitoringListener):
         self.send_custom_metrics = self.settings.get("send-custom-metrics", self.send_custom_metrics)
         self.send_custom_tables = self.settings.get("send-custom-tables", self.send_custom_tables)
         monitoring_buffer_limit = self.settings.get("monitoring-buffer-limit", 500)
-        self.monitoring_buffer = MonitoringBuffer(monitoring_buffer_limit)
+        self.monitoring_buffer = MonitoringBuffer(monitoring_buffer_limit, self.log)
         self.browser_open = self.settings.get("browser-open", self.browser_open)
         token = self.settings.get("token", "")
         if not token:
@@ -144,17 +144,10 @@ class BlazeMeterUploader(Reporter, AggregatorListener, MonitoringListener):
         self.client.log = self.log.getChild(self.__class__.__name__)
 
         if not self.client.session_id:
-            try:
-                url = self.client.start_online(self.test_id, self.sess_name)
-                self.log.info("Started data feeding: %s", url)
-                if self.browser_open in ('start', 'both'):
-                    open_browser(url)
-            except KeyboardInterrupt:
-                raise
-            except BaseException as exc:
-                self.log.debug("Exception: %s", traceback.format_exc())
-                self.log.warning("Failed to start feeding: %s", exc)
-                raise
+            url = self.client.start_online(self.test_id, self.sess_name)
+            self.log.info("Started data feeding: %s", url)
+            if self.browser_open in ('start', 'both'):
+                open_browser(url)
 
     def __get_jtls_and_more(self):
         """
@@ -416,9 +409,10 @@ class BlazeMeterUploader(Reporter, AggregatorListener, MonitoringListener):
 
 
 class MonitoringBuffer(object):
-    def __init__(self, size_limit):
+    def __init__(self, size_limit, parent_log):
         self.size_limit = size_limit
         self.data = defaultdict(OrderedDict)
+        self.log = parent_log.getChild(self.__class__.__name__)
         # data :: dict(datasource -> dict(interval -> datapoint))
         # datapoint :: dict(metric -> value)
 
@@ -437,6 +431,7 @@ class MonitoringBuffer(object):
         for source in sources:
             if len(self.data[source]) > self.size_limit:
                 self._downsample(self.data[source])
+            self.log.debug("Monitoring buffer size '%s': %s", source, len(self.data[source]))
 
     def _downsample(self, buff):
         size = 1
@@ -616,6 +611,7 @@ class BaseCloudTest(object):
     """
     :type client: BlazeMeterClient
     """
+
     def __init__(self, client, test_id, project_id, test_name, default_location, parent_log):
         self.default_test_name = "Taurus Test"
         self.client = client
@@ -984,6 +980,7 @@ class BlazeMeterClient(object):
         if method:
             req.get_method = lambda: method
 
+        start = time.time()
         response = urlopen(req, timeout=self.timeout)
 
         if checker:
@@ -993,7 +990,15 @@ class BlazeMeterClient(object):
         if not isinstance(resp, str):
             resp = resp.decode()
 
+        end = time.time()
+
         self.log.debug("Response: %s", resp[:self.logger_limit] if resp else None)
+
+        self.log.debug("Request timing: %s took %.2fs (%.2f kb sent) (%.2f kb recv)",
+                       sys._getframe(1).f_code.co_name,
+                       end - start,
+                       len(data) / 1024.0 if data else 0,
+                       len(resp) / 1024.0 if resp else 0)
         try:
             return json.loads(resp) if len(resp) else {}
         except ValueError:
@@ -1264,30 +1269,36 @@ class BlazeMeterClient(object):
         return tests['result']
 
     def __get_kpi_body(self, data_buffer, is_final):
+        # - reporting format:
+        #   {labels: <data>,    # see below
+        #    sourceID: <id of BlazeMeterClient object>,
+        #    [is_final: True]}  # for last report
+        #
+        # - elements of 'data' are described in __get_label()
+        #
+        # - elements of 'intervals' are described in __get_interval()
+        #   every interval contains info about response codes have gotten on it.
         report_items = BetterDict()
-        for dpoint in data_buffer:
-            self.first_ts = min(self.first_ts, dpoint[DataPoint.TIMESTAMP])
-            self.last_ts = max(self.last_ts, dpoint[DataPoint.TIMESTAMP])
+        if data_buffer:
+            self.first_ts = min(self.first_ts, data_buffer[0][DataPoint.TIMESTAMP])
+            self.last_ts = max(self.last_ts, data_buffer[-1][DataPoint.TIMESTAMP])
 
-            for label, kpi_set in iteritems(dpoint[DataPoint.CURRENT]):
-                report_item = report_items.get(label, self.__label_skel(label))
+            # following data is received in the cumulative way
+            for label, kpi_set in iteritems(data_buffer[-1][DataPoint.CUMULATIVE]):
+                report_item = self.__get_label(label, kpi_set)
+                self.__add_errors(report_item, kpi_set)  # 'Errors' tab
+                report_items[label] = report_item
 
-                interval_item = self.__interval_json(kpi_set, dpoint)
-                for r_code, cnt in iteritems(kpi_set[KPISet.RESP_CODES]):
-                    fails = [err['cnt'] for err in kpi_set[KPISet.ERRORS] if str(err['rc']) == r_code]
-                    interval_item['rc'].append({"n": cnt, 'f': fails, "rc": r_code})
-
-                report_item['intervals'].append(interval_item)
-
-                cumul = dpoint[DataPoint.CUMULATIVE][label]
-                report_item['n'] = cumul[KPISet.SAMPLE_COUNT]
-                report_item["summary"] = self.__summary_json(cumul)
-
-                self.__add_errors(report_item, kpi_set)
+            # fill 'Timeline Report' tab with intervals data
+            # intervals are received in the additive way
+            for dpoint in data_buffer:
+                time_stamp = dpoint[DataPoint.TIMESTAMP]
+                for label, kpi_set in iteritems(dpoint[DataPoint.CURRENT]):
+                    report_item = report_items.get(label, ValueError('Cumulative KPISet non-consistent'))
+                    report_item['intervals'].append(self.__get_interval(kpi_set, time_stamp))
 
         report_items = [report_items[key] for key in sorted(report_items.keys())]  # convert dict to list
         data = {"labels": report_items, "sourceID": id(self)}
-
         if is_final:
             data['final'] = True
 
@@ -1309,7 +1320,6 @@ class BlazeMeterClient(object):
                     'name': 'All Assertions',
                     'failures': error['cnt']
                     # TODO: "count", "errors" = ? (according do Udi's format description)
-                    # TODO: Errtype == embedded_resources ?
                 })
 
     def send_kpi_data(self, data_buffer, is_check_response=True, is_final=False):
@@ -1336,26 +1346,27 @@ class BlazeMeterClient(object):
                 self.log.info("Test was stopped through Web UI: %s", result['status'])
                 raise ManualShutdown("The test was interrupted through Web UI")
 
-    def __label_skel(self, name):
+    def __get_label(self, name, cumul):
         return {
-            "n": None,
-            "name": name if name else 'ALL',
-            "interval": 1,
-            "intervals": [],
-            "samplesNotCounted": 0,
-            "assertionsNotCounted": 0,
-            "failedEmbeddedResources": [],
-            "failedEmbeddedResourcesSpilloverCount": 0,
-            "otherErrorsCount": 0,
-            "errors": [],
-            "assertions": [],
-            "percentileHistogram": [],
-            "percentileHistogramLatency": [],
-            "percentileHistogramBytes": [],
-            "empty": False,
+            "n": cumul[KPISet.SAMPLE_COUNT],  # total count of samples
+            "name": name if name else 'ALL',  # label
+            "interval": 1,  # not used
+            "intervals": [],  # list of intervals, fill later
+            "samplesNotCounted": 0,  # not used
+            "assertionsNotCounted": 0,  # not used
+            "failedEmbeddedResources": [],  # not used
+            "failedEmbeddedResourcesSpilloverCount": 0,  # not used
+            "otherErrorsCount": 0,  # not used
+            "errors": [],  # list of errors, fill later
+            "assertions": [],  # list of assertions, fill later
+            "percentileHistogram": [],  # not used
+            "percentileHistogramLatency": [],  # not used
+            "percentileHistogramBytes": [],  # not used
+            "empty": False,  # not used
+            "summary": self.__get_summary(cumul)  # summary info
         }
 
-    def __summary_json(self, cumul):
+    def __get_summary(self, cumul):
         return {
             "first": self.first_ts,
             "last": self.last_ts,
@@ -1385,14 +1396,23 @@ class BlazeMeterClient(object):
             "otherErrorsSpillcount": 0,
         }
 
-    def __interval_json(self, item, sec):
+    def __get_interval(self, item, time_stamp):
+        #   rc_list - list of info about response codes:
+        #   {'n': <number of code encounters>,
+        #    'f': <number of failed request (e.q. important for assertions)>
+        #    'rc': <string value of response code>}
+        rc_list = []
+        for r_code, cnt in iteritems(item[KPISet.RESP_CODES]):
+            fails = [err['cnt'] for err in item[KPISet.ERRORS] if str(err['rc']) == r_code]
+            rc_list.append({"n": cnt, 'f': fails, "rc": r_code})
+
         return {
             "ec": item[KPISet.FAILURES],
-            "ts": sec[DataPoint.TIMESTAMP],
+            "ts": time_stamp,
             "na": item[KPISet.CONCURRENCY],
             "n": item[KPISet.SAMPLE_COUNT],
             "failed": item[KPISet.FAILURES],
-            "rc": [],  # filled later
+            "rc": rc_list,
             "t": {
                 "min": int(1000 * item[KPISet.PERCENTILES]["0.0"]) if "0.0" in item[KPISet.PERCENTILES] else 0,
                 "max": int(1000 * item[KPISet.PERCENTILES]["100.0"]) if "100.0" in item[KPISet.PERCENTILES] else 0,