+++ site/dat/docs/Changelog.md
@@ -2,7 +2,14 @@
 
 ## 1.7.3 <sup>next</sup>
  - fix assertion disappearance in nose_plugin
-  
+ - improve exception logging in Engine/CLI
+ - fix unicode issues in jmx2yaml
+ - make Taurus shortcut (Windows) open command line
+ - handle the absence of backtrace in RSpec
+ - jmx2yaml: use `HTTPSampler.path` even when there's no `HTTPSampler.domain` set
+ - fix cumulative part of BlazeMeter reporting
+ - remove MirrorsManager from Gatling
+
 ## 1.7.2 <sup>13 oct 2016</sup>
  - fix keep-alive processing in Gatling
  - add ability of JMeter variables usage in data-sources path options